Deletion only with valid token by the original poster

+++ backend/controllers/blogposts.js
@@ -43,6 +43,20 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: "invalid token"})
+    }
+    console.log(request.params.id)
+    const blog = await Blog.findById(request.params.id)
+    const user = await User.findById(decodedToken.id)
+
+    if ( blog.user.toString() !== user._id.toString()) {
+      return response.status(401).json({error: "Unauthorized user cannot delete posts"})
+    }
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
   } catch (exception) {
